requirements.txt
aiogram==2.17.1
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.9
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
typing-extensions==4.0.1
yarl==1.7.2
